chore(blockifier): create a file dedicated to l1 handler txs
