+++ crates/blockifier/src/transaction.rs
@@ -2,6 +2,7 @@ pub mod account_transaction;
 #[cfg(test)]
 pub mod error_format_test;
 pub mod errors;
+pub mod l1_handler_transaction;
 pub mod objects;
 #[cfg(any(feature = "testing", test))]
 pub mod test_utils;
